Minor refactoring: SimpleMarket moved to generic
