+++ dataModule/src/main/java/com/aneonex/bitcoinchecker/datamodule/model/market/CryptoCom.kt
@@ -3,8 +3,8 @@ package com.aneonex.bitcoinchecker.datamodule.model.market
 import com.aneonex.bitcoinchecker.datamodule.exceptions.MarketParseException
 import com.aneonex.bitcoinchecker.datamodule.model.CheckerInfo
 import com.aneonex.bitcoinchecker.datamodule.model.CurrencyPairInfo
-import com.aneonex.bitcoinchecker.datamodule.model.SimpleMarket
 import com.aneonex.bitcoinchecker.datamodule.model.Ticker
+import com.aneonex.bitcoinchecker.datamodule.model.market.generic.SimpleMarket
 import org.json.JSONObject
 
 class CryptoCom : SimpleMarket(
