+++ dataModule/src/main/java/com/aneonex/bitcoinchecker/datamodule/model/market/LiteBit.kt
@@ -2,9 +2,9 @@ package com.aneonex.bitcoinchecker.datamodule.model.market
 
 import com.aneonex.bitcoinchecker.datamodule.model.CheckerInfo
 import com.aneonex.bitcoinchecker.datamodule.model.CurrencyPairInfo
-import com.aneonex.bitcoinchecker.datamodule.model.SimpleMarket
 import com.aneonex.bitcoinchecker.datamodule.model.Ticker
 import com.aneonex.bitcoinchecker.datamodule.model.currency.Currency
+import com.aneonex.bitcoinchecker.datamodule.model.market.generic.SimpleMarket
 import org.json.JSONObject
 import java.util.*
